D-Roooooood.
Thanks for playing a hard D on my bad code. I'm not a great
developer but I would be 100% worse if it weren't for your
guidance! Thanks for everything, it's been real, and you're
going to kill it at Morning Consult, they're lucky to have
you. Looking forward to seeing whatever you end up doing on
this weird blue sphere -- stay in touch dude 🙂.

- Paul