SAILing away
D-Rod! You are cool as a cucumber, a fearless leader, an incredibly talented developer, and have excellent taste in EDM. Going sailing with you on the Charles in the midst of a stormy client time remains such a peaceful oasis of a memory. I hope you find ways to keep sailing in DC, and thanks so much for all you've done. You're a really rad person and a really skilled dev and we'll all miss you terribly.