Farewell, and good luck
There'll be a huge hole in the web team once you leave, but you are also leaving behind a lasting legacy on what it means to be a web lead. I wish we had more opportunities to work together, but I've definitely enjoyed the few times we had. Best of luck with your future endeavors.