Actually Knar
I’m so glad I got the chance to work with you! So many incredible memories, whether it was speaking Spanish on the ski trip or going to the movie theatre with a whole Intrepid crew to see the Armenian move The Promise. DC is gonna be amazing, good luck with the move and see you next time I’m there! We’ll miss you!

--from Knar