“Essentially” this is not a good bye, but see you later “for sure”!!!
Your physical presence will be missed soooo much, but the print you left on the team (and company) will be remembered for a loooong time! I will always think of you and your positive attitude when the shit hits the fan on the project bc you always smile and keep calm (even if on the inside this is not the case). 
On the other hand I’m very happy for you and what comes ahead of you! Wishing you lots of success at your new job, love and happiness for eternity in your marriage and lots of good and positive people around you!!!

Good luck!!!
Hugs, Svetlana